fix: avoid loosing graph state when updating inputs
The graph state was to high in the hierarky making the tabs reset
when input values was changes. This change moves the state down
one level to let it update without clearing the navigation.
this also lets the tabs remeber their position so that you can
jump back and forth between service and desktop while the
extension keeps each graph view open.